% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{compute_necessity}
\alias{compute_necessity}
\title{compute_necessity(): compute whether X was necessary for Y in the actual
world.}
\usage{
compute_necessity(x_var, y_var, causal_model, actual_world)
}
\arguments{
\item{x_var}{Character string giving the candidate cause variable.}

\item{y_var}{Character string giving the outcome variable.}

\item{causal_model}{A named list of structural functions defining
the causal model.}

\item{actual_world}{A named list giving the values of variables in
the actual world.}
}
\value{
A logical value indicating whether the candidate cause is
necessary for the outcome in the actual world.
}
\description{
We check whether intervening on X in the actual world
flips the value of Y.
}
