% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural_functions.R
\name{make_function_list}
\alias{make_function_list}
\title{Create structural functions for a causal model}
\usage{
make_function_list(vars)
}
\arguments{
\item{vars}{A named list specifying the causal model.
Elements representing structural equations should be strings, while
exogenous variables are specified by their probabilities.}
}
\value{
A named list of functions corresponding to the variables in
the causal model.
}
\description{
Converts a named causal model into a named list of R functions.
Each variable in the causal model is converted using
\code{\link[=create_structural_function]{create_structural_function()}}.
}
