% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{ns}
\alias{ns}
\title{Function computing a judgment for the NS model}
\usage{
ns(var, outcome, actual_world, d, causal_model = cm)
}
\arguments{
\item{var}{Character string giving the candidate cause variable.}

\item{outcome}{Character string giving the outcome variable.}

\item{actual_world}{A named list specifying the values of variables
in the actual world.}

\item{d}{The dataframe containing the counterfactual-based joint
distribution.}

\item{causal_model}{A named list specifying the causal model.}
}
\value{
A numeric causal judgment.
}
\description{
Function computing a judgment for the NS model
}
