% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ces.R
\name{ces}
\alias{ces}
\title{Computes a judgment with the CES model}
\usage{
ces(var1, var2, aw_values, d, p_col = "p")
}
\arguments{
\item{var1}{Character string giving the candidate cause variable.}

\item{var2}{Character string giving the outcome variable.}

\item{aw_values}{A named list specifying the values of variables
in the actual world.}

\item{d}{The dataframe containing the counterfactual-based joint distribution}

\item{p_col}{Indicates which column contains the joint probability.}
}
\value{
A numeric causal judgment.
}
\description{
Computes a judgment with the CES model
}
