% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judgment.R
\name{compute_judgment}
\alias{compute_judgment}
\title{The general causal judgment function. It is essentially a wrapper
over the ces() and ns() functions.}
\usage{
compute_judgment(var, outcome, causal_model, actual_world, model, s = 0)
}
\arguments{
\item{var}{Character string giving the candidate cause variable.}

\item{outcome}{Character string giving the outcome variable.}

\item{causal_model}{A named list specifying the causal model.}

\item{actual_world}{A named list specifying the values of variables
in the actual world. The names must match those in \code{causal_model}.}

\item{model}{Character string specifying the causal judgment model.
Currently, \code{"ces"} and \code{"ns"} are supported.}

\item{s}{A numeric: the stability parameter.}
}
\value{
A numeric causal judgment.
}
\description{
The general causal judgment function. It is essentially a wrapper
over the ces() and ns() functions.
}
