% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ns.R
\name{ns}
\alias{ns}
\title{ns(): compute a judgment with the NS model.}
\usage{
ns(var, outcome, actual_world, d, causal_model = cm)
}
\arguments{
\item{var}{Character string giving the candidate cause variable.}

\item{outcome}{Character string giving the outcome variable.}

\item{actual_world}{A named list specifying the values of variables
in the actual world.}

\item{d}{The dataframe containing the probability distribution over
counterfactual worlds.}

\item{causal_model}{A named list specifying the causal model.}
}
\value{
A numeric causal judgment.
}
\description{
This function computes a NS judgment on the basis of a probability
distribution over counterfactual worlds
}
